poi列表需要type字段支持

+++ PeachTravel/TravelPlan/PoiDetail/CityDetail/CityDetailTableViewController.m
@@ -137,7 +137,7 @@ static NSString * const reuseIdentifier = @"travelNoteCell";
     
     AccountManager *accountManager = [AccountManager shareAccountManager];
     if (accountManager.isLogin) {
-        [manager.requestSerializer setValue:[NSString stringWithFormat:@"%ld", accountManager.account.userId] forHTTPHeaderField:@"UserId"];
+        [manager.requestSerializer setValue:[NSString stringWithFormat:@"%ld", (long)accountManager.account.userId] forHTTPHeaderField:@"UserId"];
     }
     
     NSString *requsetUrl = [NSString stringWithFormat:@"%@%@", API_GET_CITYDETAIL, _cityId];
